build: change from gradle to maven
It would be nice use use Gradle for "backend-test" because it is
used in the "backend" repository. However, it seems like the
Cucumber test report are better supported with Maven. There are
plugins for this in Gradle as well but two of the most prominent
ones didn't get updates for two years.